requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
idna==3.6
inflection==0.5.1
packaging==23.2
pillow==10.2.0
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
setuptools==69.0.3
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.1.0
